requirements-dev.txt
boto3==1.34.69
pytest==8.1.1
pytest-cov==4.1.0
moto==5.0.3
pillow==10.2.0

